lua interface function
