+++ src/local_context.cpp
@@ -1,4 +1,5 @@
 #include "local_context.hpp"
+
 #include "global_context.hpp"
 #include "util/lua/lua_logging.hpp"
 
@@ -10,12 +11,14 @@ LocalContext::LocalContext(/* args */) {
     lua_logging::setup_lua_logging(lua_state);
 }
 
-LocalContext& LocalContext::get_local_context() {
+LocalContext&
+LocalContext::get_local_context() {
     auto thread_id = std::this_thread::get_id();
-    //GlobalContext& context = GlobalContext::instance();
+    // GlobalContext& context = GlobalContext::instance();
     return GlobalContext::get_local(thread_id);
 }
 
-sol::state& LocalContext::get_lua_state() {
+sol::state&
+LocalContext::get_lua_state() {
     return get_local_context().lua_state;
 }
