+++ src/local_context.hpp
@@ -1,11 +1,11 @@
-// local context 
+// local context
 // for local lua, and anything else that is local
 #pragma once
 
-//#include "logging.hpp"
-//#include "global_context.hpp"
+// #include "logging.hpp"
+// #include "global_context.hpp"
 #include <sol/sol.hpp>
-//#include <thread>
+// #include <thread>
 
 class GlobalCnotext;
 
@@ -13,9 +13,10 @@ class LocalContext {
  private:
     LocalContext();
     sol::state lua_state;
+
  public:
     static LocalContext& get_local_context();
     static sol::state& get_lua_state();
 
- friend class GlobalContext;
+    friend class GlobalContext;
 };
