+++ src/world/biome.hpp
@@ -123,10 +123,9 @@ class Biome {
 
     GrassData grass_data_;
 
-    // lua state that generates macro map
-    sol::state lua_;
+    const size_t seed_;
 
-    size_t seed_;
+    const std::filesystem::path lua_map_generator_file_;
 
  public:
     /**
@@ -275,10 +274,6 @@ class Biome {
     [[nodiscard]] std::unordered_map<ColorInt, MaterialColor>
     get_colors_inverse_map() const;
 
-    static void init_lua_state(
-        sol::state& lua, const std::filesystem::path& lua_map_generator_file
-    );
-
     static TerrainMacroMap map_generation_test(
         const std::filesystem::path& lua_map_generator_file, size_t size
     );
@@ -296,10 +291,7 @@ class Biome {
     // read data to generate the add to top after affect
     void read_add_to_top_data_(const std::vector<layer_effects_t>& biome_data);
 
-    void
-    init_lua_state_(const std::filesystem::path& lua_map_generator_file) {
-        init_lua_state(lua_, lua_map_generator_file);
-    };
+    void init_lua_state_(const std::filesystem::path& lua_map_generator_file);
 
     /**
      * @brief Load materials from json data
