+++ src/world/terrain/generation/lua_interface.hpp
@@ -1,0 +1,90 @@
+#pragma once
+#include <sol/sol.hpp>
+
+#include "noise.hpp"
+#include "worley_noise.hpp"
+
+namespace terrain {
+namespace generation {
+void init_lua_interface(sol::state& lua) {
+        lua.new_usertype<FractalNoise>(
+        "FractalNoise", sol::meta_function::construct,
+        sol::factories(
+            // FractalNoise.new(...) -- dot syntax, no "self" value
+            // passed in
+            [](int num_octaves, double persistence, int prime_index) {
+                return std::make_shared<FractalNoise>(
+                    num_octaves, persistence, prime_index
+                );
+            },
+            // FractalNoise:new(...) -- colon syntax, passes in the
+            // "self" value as first argument implicitly
+            [](sol::object, int num_octaves, double persistence, int prime_index) {
+                return std::make_shared<FractalNoise>(
+                    num_octaves, persistence, prime_index
+                );
+            }
+        ),
+        // FractalNoise(...) syntax, only
+        sol::call_constructor,
+        sol::factories([](int num_octaves, double persistence, int prime_index) {
+            return std::make_shared<FractalNoise>(
+                num_octaves, persistence, prime_index
+            );
+        }),
+        "sample", &FractalNoise::get_noise
+    );
+
+    lua.new_usertype<WorleyNoise>(
+        "WorleyNoise", sol::meta_function::construct,
+        sol::factories(
+            // WorleyNoise.new(...) -- dot syntax, no "self" value
+            // passed in
+            [](int tile_size, double radius) {
+                return std::make_shared<WorleyNoise>(tile_size, radius);
+            },
+            // WorleyNoise:new(...) -- colon syntax, passes in the
+            // "self" value as first argument implicitly
+            [](sol::object, int tile_size, double radius) {
+                return std::make_shared<WorleyNoise>(tile_size, radius);
+            }
+        ),
+        // WorleyNoise(...) syntax, only
+        sol::call_constructor, sol::factories([](int tile_size, double radius) {
+            return std::make_shared<WorleyNoise>(tile_size, radius);
+        }),
+        "sample", &WorleyNoise::get_noise
+    );
+
+    lua.new_usertype<AlternativeWorleyNoise>(
+        "AlternativeWorleyNoise", sol::meta_function::construct,
+        sol::factories(
+            // AlternativeWorleyNoise.new(...) -- dot syntax, no "self" value
+            // passed in
+            [](int tile_size, double positive_chance, double radius) {
+                return std::make_shared<AlternativeWorleyNoise>(
+                    tile_size, positive_chance, radius
+                );
+            },
+            // AlternativeWorleyNoise:new(...) -- colon syntax, passes in the
+            // "self" value as first argument implicitly
+            [](sol::object, int tile_size, double positive_chance, double radius) {
+                return std::make_shared<AlternativeWorleyNoise>(
+                    tile_size, positive_chance, radius
+                );
+            }
+        ),
+        // AlternativeWorleyNoise(...) syntax, only
+        sol::call_constructor,
+        sol::factories([](int tile_size, double positive_chance, double radius) {
+            return std::make_shared<AlternativeWorleyNoise>(
+                tile_size, positive_chance, radius
+            );
+        }),
+        "sample", &AlternativeWorleyNoise::get_noise
+    );
+
+}
+}
+}
+
